+++ packages/anon-aadhaar-react/test/hook.test.tsx
@@ -7,9 +7,10 @@ import {
   AnonAadhaarContext,
   AnonAadhaarState,
 } from '../src/hooks/useAnonAadhaar'
-import { AnonAadhaarPCDArgs, genData } from 'anon-aadhaar-pcd'
+import { AnonAadhaarPCDArgs } from 'anon-aadhaar-pcd'
 import { ArgumentTypeName } from '@pcd/pcd-types'
 import { AnonAadhaarProvider } from '../src/provider/AnonAadhaarProvider'
+import { genData } from '../../anon-aadhaar-pcd/test/utils'
 
 describe('useCountryIdentity Hook', () => {
   let testData: [bigint, bigint, bigint, bigint]
